the resource calls Helm but then helm overwrites the previous version

+++ src/filesystem/gitclone.go
@@ -7,7 +7,9 @@ import (
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing/transport/ssh"
 
-	"github.com/lithammer/shortuuid/v3"
+	"crypto/sha1"
+	"encoding/base64"
+
 	"github.com/rs/zerolog/log"
 )
 
@@ -18,7 +20,11 @@ func Clone2Tmp(url string) (path string, err error) {
 		return "", err
 	}
 
-	path = filepath.Join(os.TempDir(), "deployto-gitclones", shortuuid.New())
+	hasher := sha1.New()
+	hasher.Write([]byte(url))
+	urlsha := base64.URLEncoding.EncodeToString(hasher.Sum(nil))
+
+	path = filepath.Join(os.TempDir(), "deployto-gitclones", urlsha)
 	err = os.MkdirAll(path, os.ModePerm)
 	if err != nil {
 		log.Error().Err(err).Msg("Can't make tmp dir for git clone")
@@ -31,12 +37,30 @@ func Clone2Tmp(url string) (path string, err error) {
 		Auth:              authMethod,
 	})
 	if err != nil {
-		log.Error().Err(err).Str("url", url).Msg("Can't clone git")
-		extraError := os.RemoveAll(path)
-		if extraError != nil {
-			log.Error().Err(extraError).Str("path", url).Msg("Can't remove tmp path")
+		if err.Error() != "repository already exists" {
+			log.Error().Err(err).Str("url", url).Msg("Can't clone git")
+			extraError := os.RemoveAll(path)
+			if extraError != nil {
+				log.Error().Err(extraError).Str("path", url).Msg("Can't remove tmp path")
+			}
+			return "", err
 		}
-		return "", err
+		// git pull
+		r, err := git.PlainOpen(path)
+		if err != nil {
+			log.Error().Err(err).Str("path", path).Msg("git.PlainOpen error")
+			return "", nil
+		}
+		w, err := r.Worktree()
+		if err != nil {
+			log.Error().Err(err).Str("path", path).Msg("git get worktree error")
+			return "", nil
+		}
+		err = w.Pull(&git.PullOptions{RemoteName: "origin"})
+		if err != nil {
+			log.Error().Err(err).Str("path", path).Msg("git pull error")
+			return "", nil
+		}
 	}
 	return path, err
 }
